make jsonserializable way
