+++ example/lib/json_serializable/base_error.dart
@@ -1,0 +1,8 @@
+
+
+class BaseError {
+  final int? errorCode;
+  final String? message;
+
+  BaseError([this.errorCode, this.message = "something went wrong!"]);
+}
