+++ example/lib/json_serializable/error_code.dart
@@ -1,0 +1,11 @@
+
+
+class ErrorCode {
+  static int notFound = 101;
+  static int unexpected = 102;
+  static int cancel = 103;
+  static int connectionTimeOut = 104;
+  static int defaultError = 105;
+  static int sendTimeout = 106;
+}
+
