+++ example/lib/json_serializable/model/post.dart
@@ -1,0 +1,28 @@
+
+import 'package:viva_network_call/viva_network_call.dart';
+
+class Post extends Serializable {
+  late int userId;
+  int? id;
+  String? title;
+  String? body;
+
+  Post({required this.userId, this.id, this.title, this.body});
+
+  Post.fromJson(Map<String, dynamic> json) {
+    userId = json['userId'];
+    id = json['id'];
+    title = json['title'];
+    body = json['body'];
+  }
+
+  @override
+  Map<String, dynamic> toJson() {
+    final Map<String, dynamic> data = new Map<String, dynamic>();
+    data['userd'] = this.userId;
+    data['id'] = this.id;
+    data['title'] = this.title;
+    data['body'] = this.body;
+    return data;
+  }
+}
